v0.2.8 - finegui now uses floating point for size and position manipulation, allowing much better precision - added mouse scroll event to finegui

+++ include/ffw/graphics/font.h
@@ -31,8 +31,8 @@ namespace ffw {
 
 		virtual void destroy() = 0;
 		virtual const Char& getChar(wchar_t chr) const = 0;
-		virtual ffw::Vec2i getStringSize(const std::string& str, float lineHeight = 1.25f) const;
-		virtual ffw::Vec2i getStringSize(const std::wstring& str, float lineHeight = 1.25f) const;
+		virtual ffw::Vec2f getStringSize(const std::string& str, float lineHeight = 1.25f) const;
+		virtual ffw::Vec2f getStringSize(const std::wstring& str, float lineHeight = 1.25f) const;
 
 		inline int getSizePt() const {
 			return sizePoints;
@@ -57,7 +57,7 @@ namespace ffw {
 		}
     protected:
 		template<class T>
-		ffw::Vec2i getStringSizeFunc(const std::basic_string<T>& str, float lineHeight) const;
+		ffw::Vec2f getStringSizeFunc(const std::basic_string<T>& str, float lineHeight) const;
         ffw::Texture2D texture;
 		int sizePoints;
 		int sizePixels;
